like logic fix, profile pic ui bug fix, react query changes to how likes are processed locally
