+++ src/app/components/ReviewCard.tsx
@@ -1,28 +1,28 @@
 "use client";
-import Image, { StaticImageData } from "next/image";
-import { RockBadge } from "../icons/stamps";
+import Image from "next/image";
 import { Icons } from "../icons/icons";
 import { useState } from "react";
-import { likeReview } from "../actions/reviews_service";
+import { toggleLike } from "../actions/reviews_service";
 import { Review } from "../types/types";
-import { useLikeReview } from "../hooks";
+import { useToggleReview } from "../hooks";
 import { genreBadge } from "../utils/calculations";
+import { useQueryClient } from "@tanstack/react-query";
 type ReviewCardProps = {
   avatar?: string;
   review: Review;
-  setReviews: React.Dispatch<React.SetStateAction<Review[]>>;
   username: string;
 };
 
 export default function ReviewCard({
   review,
-  setReviews,
   avatar,
   username,
 }: ReviewCardProps) {
-  const { mutate: likeReviewMutate } = useLikeReview(username);
+  const queryClient = useQueryClient();
+  const { mutate: toggleReviewMutate } = useToggleReview(username);
   const [likingReviews, setLikingReviews] = useState<Set<number>>(new Set());
-
+  const [liked, setLiked] = useState(review.is_liked_by_user);
+  const [likeCount, setLikeCount] = useState(review.likes_count);
   const formattedDate = new Date(review.created_at).toLocaleDateString(
     "en-US",
     {
@@ -31,30 +31,35 @@ export default function ReviewCard({
       day: "numeric",
     }
   );
-  const handleLikeReview = (reviewId: number, currentlyLiked: boolean) => {
+  const handletoggleReview = (reviewId: number, currentlyLiked: boolean) => {
     if (likingReviews.has(reviewId)) return;
-
     setLikingReviews((prev) => new Set(prev).add(reviewId));
+    setLiked(!currentlyLiked);
+    setLikeCount((prev) => (currentlyLiked ? prev - 1 : prev + 1));
 
-    likeReviewMutate(reviewId, {
+    toggleReviewMutate(reviewId, {
       onSuccess: () => {
-        setReviews((prev) =>
-          prev.map((review) =>
-            review.id === reviewId
-              ? {
-                  ...review,
-                  is_liked_by_user: !currentlyLiked,
-                  likes_count: currentlyLiked
-                    ? review.likes_count - 1
-                    : review.likes_count + 1,
-                }
-              : review
-          )
+        queryClient.setQueryData(
+          ["reviews", username],
+          (prev: Review[] | undefined) =>
+            prev?.map((r) =>
+              r.id === reviewId
+                ? {
+                    ...r,
+                    is_liked_by_user: !currentlyLiked,
+                    likes_count: currentlyLiked
+                      ? r.likes_count - 1
+                      : r.likes_count + 1,
+                  }
+                : r
+            )
         );
       },
       onError: (error) => {
         //do something else
         console.error(error);
+        setLiked(currentlyLiked);
+        setLikeCount((prev) => (currentlyLiked ? prev + 1 : prev - 1));
         alert("Error updating like status");
       },
       onSettled: () => {
@@ -123,7 +128,7 @@ export default function ReviewCard({
 
           <button
             disabled={likingReviews.has(review.id)}
-            onClick={() => handleLikeReview(review.id, review.is_liked_by_user)}
+            onClick={() => handletoggleReview(review.id, liked)}
             className={`flex items-center justify-center gap-1 border border-black rounded-full bg-[#f4f4f4] text-sm text-black w-12 h-7 transition-opacity duration-200 ${
               likingReviews.has(review.id)
                 ? "opacity-50 cursor-not-allowed"
@@ -131,17 +136,13 @@ export default function ReviewCard({
             }`}
           >
             <Image
-              src={
-                review.is_liked_by_user ? Icons.likedHeart : Icons.unlikedHeart
-              }
+              src={liked ? Icons.likedHeart : Icons.unlikedHeart}
               alt="Favorite Icon"
               width={12}
               height={12}
               className="object-contain"
             />
-            <span className="text-[13px] font-medium">
-              {review.likes_count}
-            </span>
+            <span className="text-[13px] font-medium">{likeCount}</span>
           </button>
         </div>
       </div>
